face/requirements.txt
keras-facenet
tensorflow==2.10
numpy
pillow
scikit-learn
joblib
streamlit
mtcnn
matplotlib
seaborn
opencv-python